--- src/containers/index.ts
@@ -1,10 +1,0 @@
-/*
- * Created on Wed Nov 22 2023
- *
- * Copyright (c) 2023 Your Company
- */
-
-export * as Login from "./Login";
-export * as Club from "./Club";
-export * as Board from "./Board";
-export * as Home from "./Home";
