+++ src/pages/login/index.tsx
@@ -6,7 +6,7 @@
 
 import Head from "next/head";
 
-import { Login } from "@/containers";
+import * as Login from "@/containers/Login";
 
 export default function LoginPage() {
   return (
